+++ packages/dataviews/CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### Breaking Changes
 
-- `onSelectionChange` has been renamed to `onChangeSelection`.
+- `onSelectionChange` prop has been renamed to `onChangeSelection` and its argument has been updated to be a list of ids.
+- `setSelection` prop has been removed. Please use `onChangeSelection` instead.
 
 ## 3.0.0 (2024-07-10)